refactor: Move logic to MapboxMap file.
