+++ packages/map-template/src/components/MapTemplate/MapTemplate.jsx
@@ -52,7 +52,6 @@ import LegendDialog from '../LegendDialog/LegendDialog.jsx';
 import isLegendDialogVisibleState from '../../atoms/isLegendDialogVisibleState.js';
 import searchAllVenuesState from '../../atoms/searchAllVenues.js';
 import currentVenueNameState from '../../atoms/currentVenueNameState.js';
-import VisibilitySwitch from '../VisibilitySwitch/VisibilitySwitch.jsx';
 
 // Define the Custom Elements from our components package.
 defineCustomElements();
@@ -129,7 +128,7 @@ function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, p
     const [initialFilteredLocations, setInitialFilteredLocations] = useState();
 
     const [appConfig, setAppConfig] = useState();
-    const [solution, setSolution] = useRecoilState(solutionState);
+    const [, setSolution] = useRecoilState(solutionState);
 
     const [, setTileStyle] = useRecoilState(tileStyleState);
     const [, setStartZoomLevel] = useRecoilState(startZoomLevelState);
@@ -155,8 +154,6 @@ function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, p
 
     // The reset count is used to add a new key to the sidebar or bottomsheet, forcing it to re-render from scratch when resetting the Map Template.
     const [resetCount, setResetCount] = useState(0);
-
-    const [showVisibilitySwitch, setShowVisibilitySwitch] = useState(false);
 
     /**
      * Ensure that MapsIndoors Web SDK is available.
@@ -519,23 +516,6 @@ function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, p
         }
     }, [currentVenueName, mapsindoorsSDKAvailable]);
 
-    /*
-     * React on changes to the solution.
-     * Decide whether to show the VisibilitySwitch component depending on the modules enabled on the solution.
-     */
-    useEffect(() => {
-        if (solution) {
-            const isMapboxModuleEnabled = solution.modules.map(module => module.toLowerCase()).includes('mapbox');
-            const is3DWallsModuleEnabled = solution.modules.map(module => module.toLowerCase()).includes('3dwalls');
-            // The module floorplan refers to 2D Walls 
-            const is2DWallsModuleEnabled = solution.modules.map(module => module.toLowerCase()).includes('floorplan');
-
-            if (isMapboxModuleEnabled && is3DWallsModuleEnabled && is2DWallsModuleEnabled) {
-                setShowVisibilitySwitch(true);
-            }
-        }
-    }, [solution]);
-
     /**
      * When venue is fitted while initializing the data,
      * set map to be ready and get the venue categories.
@@ -624,7 +604,6 @@ function MapTemplate({ apiKey, gmApiKey, mapboxAccessToken, venue, locationId, p
             onClose={() => goBack()}
             active={currentAppView === appStates.VENUE_SELECTOR}
         />}
-        {showVisibilitySwitch && <VisibilitySwitch />}
         {showQRCodeDialog && <QRCodeDialog />}
         {showLegendDialog && <LegendDialog />}
         {isMapPositionKnown &&
